ot_app_pair-refactor: rewrite pairing task to use zero-copy buffer API
- Completely refactored `otapp_pair_task` and `xQueueReceive` logic to improve stability and flow.
- Updated `otapp_pair_responseHandlerUriWellKnown` to utilize the new shared buffer mechanism.
- Integrated `otapp_buf_getWriteOnly_ptr`: replaced temporary local buffers with direct, thread-safe access to the global buffer (Zero-Copy), reducing stack usage and `memcpy` operations.

+++ app/ot_app/inc/ot_app_pair.h
@@ -251,7 +251,7 @@ int8_t otapp_pair_addToQueue(otapp_pair_queueItem_t *queueItem);
  */
 otapp_pair_resUrisParseData_t *otapp_pair_uriParseMessage(uint8_t *inBuffer, uint16_t inBufferSize, int8_t *result, uint16_t *outParsedDataSize);
 
-uint16_t otapp_pair_uriParseMessageCalculateBufSize(uint8_t aMessagePayloadSize);
+uint16_t otapp_pair_uriParseMessageCalculateBufSize(uint16_t aMessagePayloadSize);
 
 /**
  * @brief
